+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.9.0
 
-* Added eDoc url to Report.
+* Added eDoc url to Report and System.
 
 ## 1.8.0